requirements（更新于20230201较全，去掉了pytorch，其他的都有）.txt
certifi
cycler
Cython
easydict
kiwisolver
llvmlite
matplotlib
numba
numpy
opencv-python
Pillow
progress
pycocotools
pyparsing
python-dateutil
scipy
six
torchsummary
tqdm